refactor: extract history storage into separate file
