+++ trin-storage/src/versioned/legacy_history/mod.rs
@@ -1,0 +1,4 @@
+mod sql;
+mod store;
+
+pub use store::LegacyHistoryStore;
